requirements.txt
Flask
numpy
opencv_python
Pillow
PyAutoGUI
Werkzeug